+++ cloud/acp-admin-cloud-dependencies/src/main/kotlin/pers/acp/admin/common/vo/InnerInfoVo.kt
@@ -7,8 +7,10 @@ import io.swagger.annotations.ApiModelProperty
  * @author zhang by 11/01/2019
  * @since JDK 11
  */
-@ApiModel("响应信息")
-data class InfoVo(
+@ApiModel("内部响应信息")
+data class InnerInfoVo(
+    @ApiModelProperty(value = "是否成功")
+    var success: Boolean = true,
     @ApiModelProperty(value = "响应信息内容")
     var message: String? = null
 )