+++ doc/version_history.md
@@ -1,4 +1,11 @@
 ## 版本更新记录
+##### v5.1.2
+- Global
+  - [Upgrade] 升级依赖
+- acp-admin-cloud-dependencies
+  - [Upgrade] 优化Feign客户端写法
+- oauth-server
+  - [Upgrade] 增加禁用用户内部接口
 ##### v5.1.1
 - Global
   - [Upgrade] hikari连接池配置中增加connection-timeout
@@ -7,8 +14,11 @@
     - kotlin 1.5.31
     - Spring Boot 2.5.5
     - kotlin coroutines 1.5.2
+- acp-admin-cloud-dependencies
+  - [Upgrade] 优化Feign客户端写法
 - oauth-server
   - [Upgrade] 增加用户密码加解密工具类
+  - [Upgrade] 增加禁用用户内部接口
 ##### v5.1.0
 > - Global
 >   - [Upgrade] 升级 Gradle 至 7.2